Added: Logic for admin and normal user

+++ src/main/java/com/example/bookstore/service/BookService.java
@@ -28,7 +28,7 @@ public class BookService implements IBookService {
     public Book addBook(String token, BookDTO bookDTO) {
         Book book = new Book(bookDTO);
         UserData userData = iUserService.getUserById((token));
-        if (userData != null) {
+        if (userData != null && userData.isAdmin()) {
             emailService.sendEmail(userData.getEmail(), "Book Created successfully!!", "User " + userData.getFirstName() + " has added new book successfully " + book + ".");
             return bookRepository.save(book);
         } else throw new CustomException("Please pass valid user token to add book");
@@ -45,8 +45,9 @@ public class BookService implements IBookService {
         } else return bookRepository.findAll();
     }
 
-    public String deleteById(int id) {
-        if (bookRepository.findById(id).isPresent()) {
+    public String deleteById(int id, String token) {
+        UserData userData = iUserService.getUserById((token));
+        if (bookRepository.findById(id).isPresent() && userData.isAdmin()) {
             bookRepository.deleteById(id);
             return "Book with ID: " + id + " is Deleted Successfully!!";
         } else throw new CustomException("No book matches with the given ID");
@@ -54,7 +55,7 @@ public class BookService implements IBookService {
 
     public Book updateById(int id, BookDTO bookDTO, String token) {
         UserData userData = iUserService.getUserById((token));
-        if (bookRepository.findById(id).isPresent() && userData != null) {
+        if (bookRepository.findById(id).isPresent() && userData != null && userData.isAdmin()) {
             Book book = new Book(id, bookDTO);
             emailService.sendEmail(userData.getEmail(), "Book Updated successfully!!", "User " + userData.getFirstName() + " has updated book successfully " + book + ".");
             return bookRepository.save(book);
@@ -103,7 +104,7 @@ public class BookService implements IBookService {
 
     public Book updateQuantityById(int id, int quantity, String token) throws CustomException {
         UserData userData = iUserService.getUserById((token));
-        if (bookRepository.findById(id).isPresent() && userData != null) {
+        if (bookRepository.findById(id).isPresent() && userData != null && userData.isAdmin()) {
             Book book = this.getById(id);
             book.setQuantity(quantity);
             return bookRepository.save(book);
